requirements.txt
obspy==1.4.0
matplotlib==3.5.2
Pillow==9.2.0
requests==2.28.1
Jinja2==3.1.2
